+++ packages/app-deploy-ops/vite/vite.config.js
@@ -9,14 +9,20 @@ import {fileURLToPath} from 'url'
 
 const myPath = dirname(fileURLToPath(import.meta.url));
 
-import { manualChunks } from '../manualChunks.js'
+import visualizer from 'rollup-plugin-visualizer'
+
+import { manualChunks } from '../common/manualChunks.js'
+
+const createStats = true;
 
 export default {
-  /*resolve: {
-    dedupe: [
+  resolve: {
+    /*dedupe: [
       'tslib'   // does not seem to help
-    ]
-  },*/
+    ]*/
+    // For dear Firebase
+    mainFields: ["esm2017", "module"]
+  },
 
   define: {     // "statically replaced" for production
     "_OPS_VERSION": "\"0.0.0\""
@@ -33,7 +39,16 @@ export default {
     //polyfillDynamicImport: false
 
     rollupOptions: {
-      output: { manualChunks }
+      output: { manualChunks },
+
+      plugins: [
+        createStats && visualizer({
+          sourcemap: true,
+          template: 'sunburst',
+          brotliSize: true,
+          filename: 'stats.vite.html'
+        })
+      ]
     }
   }
 }